requirements.txt
PyQt5
pandas
mysql-connector-python
pyodbc
openpyxl